🚧 スキャナーを起動する前にワンクッション挟むように
